requirements.txt
arxiv
requests
beautifulsoup4
transformers
torch
flask
python-dotenv
newspaper3k
openai
sqlalchemy
schedule 